fix: deploys always being triggered from scratch
Due to a change in either GitHub Actions or the scp version used on Combell servers, the read-server-revision step was unable to determine the correct revision. The scp command would result in the following exception:
```bash
scp: local ftruncate "/dev/stdout": Invalid argument
scp: write local "/dev/stdout": Illegal seek
```
Since the command uses `2>/dev/null`, the ouput of this exception would be transformed into null. Later on, this revision is used to determine the diff of what should be deployed. If the revision is empty, a scratch deploy will be performed. We now use ssh and use the `cat` method, which should be available on all Linux systems.